Tablet configuration - Padding dimens - Map button move to right on tablet mode
