+++ app/src/main/res/layout-xlarge/activity_main.xml
@@ -10,14 +10,13 @@
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
+        android:layout_height="wrap_content"
+        android:theme="@style/AppTheme.AppBarOverlay">
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:popupTheme="@style/AppTheme.PopupOverlay" />
-
     </com.google.android.material.appbar.AppBarLayout>
 
     <LinearLayout
@@ -40,5 +39,14 @@
             android:layout_height="match_parent"
             android:layout_weight="2" />
     </LinearLayout>
+
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/mapFab"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_margin="16dp"
+        app:srcCompat="@drawable/ic_map_black_24dp"
+        android:tint="@color/colorPrimaryDark"
+        android:layout_gravity="bottom|start"/>
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
