+++ app/src/main/res/layout/activity_search.xml
@@ -11,7 +11,10 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="vertical"
-        android:padding="16dp">
+        android:paddingLeft="@dimen/property_details_horizontal"
+        android:paddingRight="@dimen/property_details_horizontal"
+        android:paddingTop="@dimen/property_details_vertical"
+        android:paddingBottom="@dimen/property_details_vertical">
 
         <TextView
             android:id="@+id/textView3"
