+++ app/src/main/res/values-large/dimens.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="property_details_horizontal">32dp</dimen>
+    <dimen name="property_details_vertical">16dp</dimen>
+</resources>
